+++ packages/backend/src/application/Repository.ts
@@ -1,5 +1,6 @@
 export interface Repository<T> {
-  // exists(t: T): Promise<boolean>
-  // delete(t: T): Promise<any>
-  // save(t: T): Promise<any>
+  create(entity: never): Promise<T>
+  findById(id: string): Promise<T>
+  update(entity: never): Promise<T>
+  deleteById(id: string): Promise<T>
 }
