+++ packages/backend/src/modules/attachment/attachment.repository.ts
@@ -1,16 +1,17 @@
 import fsPromises from 'fs/promises'
-import { CreateAttachmentDTO } from 'shared'
+import { CreateAttachmentCommand } from 'shared'
 import { Attachment } from './attachment'
 import AttachmentModel from './attachment.model'
+import { Repository } from '../../application/Repository'
 
-export class AttachmentRepository {
+export class AttachmentRepository implements Repository<Attachment> {
   private readonly _attachmentModel
 
   constructor(attachmentModel: any) {
     this._attachmentModel = attachmentModel
   }
 
-  async createAttachment(createAttachment: CreateAttachmentDTO): Promise<Attachment> {
+  async create(createAttachment: CreateAttachmentCommand): Promise<Attachment> {
     const savedAttachment = await this._attachmentModel.create({
       taskId: createAttachment.taskId,
       fileName: createAttachment.file.originalname,
@@ -20,16 +21,21 @@ export class AttachmentRepository {
     return savedAttachment
   }
 
+  async update(createAttachment: Attachment): Promise<Attachment> {
+    throw new Error('Method not implemented.')
+  }
+
   async findById(id: string): Promise<Attachment> {
     const attachment: Attachment = await this._attachmentModel.findById(id)
     return attachment
   }
 
-  async deleteById(id: string): Promise<Attachment | null> {
+  async deleteById(id: string): Promise<Attachment> {
     const file = await this._attachmentModel.findById(id)
-    if (!file) {
-      return null
-    }
+    // TODO: this method shouldn't return null
+    // if (!file) {
+    //   return null
+    // }
     await fsPromises.unlink(`uploads/${file.fileNameHash}`)
     const deletedAttachment = await this._attachmentModel.findByIdAndDelete(id)
     return deletedAttachment
