+++ packages/shared/src/api/command/createAttachment.command.ts
@@ -1,0 +1,6 @@
+interface CreateAttachmentCommand {
+  taskId: string
+  file: File
+}
+
+export default CreateAttachmentCommand
