--- packages/shared/src/api/dto/attachment/createAttachment.dto.ts
@@ -1,6 +1,0 @@
-interface CreateAttachmentDTO {
-  taskId: string
-  file: File
-}
-
-export default CreateAttachmentDTO
