+++ packages/shared/src/api/index.ts
@@ -2,5 +2,5 @@ export type { default as CommentTypeDTO } from './dto/comment.dto'
 export type { default as TaskDTO } from './dto/task.dto'
 export type { default as UserTypeDTO } from './dto/user.dto'
 export type { default as CreateTaskCommandDTO } from './dto/CreateTaskCommand.dto'
-export type { default as AttachmentDTO } from './dto/attachment/attachment.dto'
-export type { default as CreateAttachmentDTO } from './dto/attachment/createAttachment.dto'
+export type { default as AttachmentDTO } from './dto/attachment.dto'
+export type { default as CreateAttachmentCommand } from './command/createAttachment.command'
